get resource links via secure connection
Current implementation of SDK expected that cloud resource is visible under unsecure connection. But according to Device2Cloud spec this is not true:
The CoAPCloudConf Resource shall expose only secure Endpoints (e.g. CoAPS); see the ISO/IEC 30118-1:2018, clause 10.
You have to be secure to talk to it. The only time we allow unsecure in any circumstance is initial MCAST discovery.