Fix for docker containers (#146)
* separate messageIDs for UDP4/UDP6 multicast messages

We need to separate messageIDs for upd4 and udp6, because if any docker container has isolated network
iotivity-lite gets error EINVAL(22) for sendmsg with UDP6 for some interfaces. If it happens, the device is
not discovered and msgid is cached so all other multicast messages from another interfaces are dropped for deduplication.