Add WithLinkNotFoundCallback for non-discoverable resources
When a resource is non-discoverable via any link or collection
we can utilize the WithLinkNotFoundCallback option and provide
the link by other means. This addition is required when you need
to provide thousands of resources and making them discoverable
or addiding them to some collection would result in an unusable
device due to resources being too big for efectively processing
them.